requirements.txt
pytest==8.2.0
playwright==1.44.0
pytest-playwright==0.5.0
pytest-html==4.1.1
pytest-bdd==8.1.0
pytest-xdist==3.6.1
google-genai==1.3.0
pillow
locust==2.27.0
axe-playwright-python==0.1.3
allure-pytest==2.13.5
